Optimize feature saving runtime
